rename directries
