+++ src/app/infrastructure/mysqlhandler.go
@@ -2,7 +2,7 @@ package infrastructure
 
 import (
 	"database/sql"
-	"interfaces/database/rdb"
+	"interfaces/repository/rdb"
 
 	_ "github.com/go-sql-driver/mysql"
 )
@@ -14,7 +14,6 @@ type SqlHandler struct {
 func NewSqlHandler() rdb.SqlHandler {
 	c := NewDBConfig()
 	dbPath := c.User + ":" + c.Password + "@tcp(" + c.Container + ":" + c.Port + ")/" + c.Database + "?parseTime=true"
-	// dbPath := "aapaca_user:passwd@tcp(192.168.99.100:3306)/aapaca?parseTime=true"
 	conn, err := sql.Open("mysql", dbPath)
 	if err != nil {
 		panic(err)
