+++ src/app/interfaces/controller/artist_controller_test.go
@@ -1,0 +1,1 @@
+package controller
