+++ src/app/interfaces/controller/error.go
@@ -1,4 +1,4 @@
-package controllers
+package controller
 
 type Error struct {
 	Message string
