+++ src/app/interfaces/repository/album_repository.go
@@ -1,10 +1,10 @@
-package database
+package repository
 
 import (
 	"database/sql"
 	"domain"
 	"errors"
-	"interfaces/database/rdb"
+	"interfaces/repository/rdb"
 )
 
 type AlbumRepository struct {
