+++ src/app/test/Tools.go
@@ -1,10 +1,13 @@
-package interfaces
+package test
 
 import (
 	"domain"
+	"net/http"
+	"net/http/httptest"
 	"sort"
 	"testing"
 
+	"github.com/labstack/echo"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -23,4 +26,15 @@ func AssertCredits(t *testing.T, expected []domain.Credit, got []domain.Credit) 
 		assert.Equal(t, expected[i].Artist, got[i].Artist)
 		assert.ElementsMatch(t, expected[i].Parts, got[i].Parts)
 	}
+}
+
+func CreateContextInstance(path string, paramName string, paramValue string) (*httptest.ResponseRecorder, echo.Context) {
+	e := echo.New()
+	req := httptest.NewRequest(http.MethodGet, "/", nil)
+	rec := httptest.NewRecorder()
+	c := e.NewContext(req, rec)
+	c.SetPath(path)
+	c.SetParamNames(paramName)
+	c.SetParamValues(paramValue)
+	return rec, c
 }
